[OPENENGSB-2728] Flat structure for example domain
