+++ domain/example/src/main/java/org/openengsb/domain/example/ExampleDomain.java
@@ -19,9 +19,6 @@ package org.openengsb.domain.example;
 
 import org.openengsb.core.api.Domain;
 import org.openengsb.core.api.Raises;
-import org.openengsb.domain.example.event.LogEvent;
-import org.openengsb.domain.example.model.ExampleRequestModel;
-import org.openengsb.domain.example.model.ExampleResponseModel;
 
 
 // @extract-start ExampleDomain
