+++ domain/example/src/main/java/org/openengsb/domain/example/ExampleDomainEvents.java
@@ -18,7 +18,6 @@
 package org.openengsb.domain.example;
 
 import org.openengsb.core.api.DomainEvents;
-import org.openengsb.domain.example.event.LogEvent;
 
 // @extract-start ExampleDomainEvents
 public interface ExampleDomainEvents extends DomainEvents {
