+++ itests/src/test/java/org/openengsb/itests/exam/WorkflowIT.java
@@ -36,9 +36,9 @@ import org.openengsb.core.api.workflow.model.RuleBaseElementType;
 import org.openengsb.core.common.AbstractOpenEngSBService;
 import org.openengsb.core.common.util.ModelUtils;
 import org.openengsb.domain.example.ExampleDomain;
-import org.openengsb.domain.example.event.LogEvent;
-import org.openengsb.domain.example.model.ExampleRequestModel;
-import org.openengsb.domain.example.model.ExampleResponseModel;
+import org.openengsb.domain.example.ExampleRequestModel;
+import org.openengsb.domain.example.ExampleResponseModel;
+import org.openengsb.domain.example.LogEvent;
 import org.openengsb.itests.util.AbstractPreConfiguredExamTestHelper;
 import org.ops4j.pax.exam.junit.ExamReactorStrategy;
 import org.ops4j.pax.exam.junit.JUnit4TestRunner;
