+++ itests/src/test/java/org/openengsb/itests/remoteclient/ExampleConnector.java
@@ -22,9 +22,9 @@ import org.openengsb.core.api.Connector;
 import org.openengsb.core.common.AbstractOpenEngSBConnectorService;
 import org.openengsb.core.common.util.ModelUtils;
 import org.openengsb.domain.example.ExampleDomain;
-import org.openengsb.domain.example.event.LogEvent;
-import org.openengsb.domain.example.model.ExampleRequestModel;
-import org.openengsb.domain.example.model.ExampleResponseModel;
+import org.openengsb.domain.example.ExampleRequestModel;
+import org.openengsb.domain.example.ExampleResponseModel;
+import org.openengsb.domain.example.LogEvent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
